+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - feat: add `NesIcons.redo`
 - feat: add `customExtensions`
 - feat: add keyboard support 
+- feat: add `NesIcons.unpressedButton`
+- feat: add `NesIcons.pressedButton`
+- feat: add `NesKeyButton`
 
 # 0.1.0